+++ src/BarcodeServiceProvider.php
@@ -19,6 +19,12 @@ class BarcodeServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                Console\GenerateBarcodeCommand::class,
+            ]);
+        }
+
         $this->publishes([
             __DIR__.'/../config/barcode.php' => config_path('barcode.php'),
         ], 'barcode-config');